made some changes
move bitstuffing programs into its own dir
there was an spelling mistake in program file
formated the files
